Barrier Synchronizat, Fix up MIProtocol and CacheCoherenceBus, restore full test
Modified Simulator to make a worker thread for each component and synchronize
them with barriers, rather than forking new threads each phase and joining
on their results. The new AsyncComponentInterface lets components directly
recieve the barriers and join on them themselves, which may let us implement
blocking functions like "send BusMessage and wait for responses".

CacheCoherence bus was recently modified to clone a memory request,
but was still spinning on the original message. Cloning was a bad
idea, as it either leaves the invalidation up to the client and
so behaves rather differently from the MemoryInterface handling,
or it means the bus marks the event and we have all the same problems
with protocols that pass in a Processor's request directly and don't
want it marked by the higher level memory.

Added logging to CacheCoherence bus, recording the current state
each tick plus information on transitions. Having Simulator.logEvent
print events as they happen is quite helpful with these new
records.

MIProtocol was broken in several ways, like always running the code for
handling a client request when it gets a memory response, even if it
made the request to writeback an evicted line, not correctly marking
the client's request finished between filling it and losing it's
reference, forgetting to null out the request reference which would
prevent it from moving on.

Restored TestCacheProtocol.java, which had been stripped down to
one core and one request to help narrow down some earlier NPE's
fixed in the preceeding commits. Removed the space after Ox when
printing the loaded result.