New state-based protocol interface
The idea is a cache controller moves through a series of states as events arrive.
The CacheControllerState describes these events. Some can be temporarily deferred
by returning null, else return a StateAnd value, which includes a response and
a next state. CacheController connects a client with this interface up to the old
CacheCoherenceBus. For the future, maybe it would be simpler if the bus worked
with this interface more directly?

The MESIProtocol directory includes a code working towards an implementation of
MESI in this interface. The MESICacheController holds common state, most notably
the LRUEvictHashTable that backs the cache. MESICacheControllerState is a parent
of the states. The controller runs in four states - in NotReady it has no
outstanding client request, and participates in other transactions and possibly
serves client requests that don't require bus access. In Ready there is an
outstanding client request it needs to hit the bus to handle. It participates
in other transactions while trying to grab the bus (ignoring further client
requests). Running operates the bus transaction. Done offers the completed
transaction back, and participates in other transactions, transitioning into
NotReady when the result is read.