Fixed several bugs in the CacheController, added logging, MESI passes TestCacheProtocol
CacheController was handling receiveBusMessage incorrectly -
it didn't let the clients that were not running the transaction
ignore the reply they saw by returning null. (Maybe they should
have to explicitly accept the notification and return a null value?).

Moved logging out of the protocol states into a NoisyCacheControllerState,
which wraps any CacheControllerState, echoing the method and arguments
before and after and the result after a call, and also wraps any
new state we transition to with another NoisyCacheControllerState.
So, to turn on extensive logging just set the initial state of the cache
controller to be one of these things around the usual initial state.