CacheController was using an ArrayDeque which was not actually safe!
The problem is the guarantee on MemoryInterface, as implemented by the
original MainMemory, is that enequeueMemoryInstruction should only
be called during the runClock phase (and at most once per runClock
if you want to preserve determinism). Instead, the CacheController
read from the queue during runClock and expected enqueue to only
be invoked during runPrep, so it used the non threadsafe ArrayDeque.

This meant memory requests from a client could be dropped!
A processor waits until the instruction is done, so this would
kill a processor. Especially bad if it held a lock, or was
central to a test.