Fixed a major bug with writeback handling, which could have caused lost writes.
A cache in the ready state waiting to send a writeback would still send that
writeback message when it gets the bus, even if other caches that beat it to
the bus sent events involving that line while it was blocked, such as a load
of that line. Also, such a get should have been handled by a dirty ack of
the data, but because the line was removed at eviction time our waiting
cache wouldn't respond to another cache's read miss, and the other cache
would go on to get stale data from memory, probably in an exclusive state.

This was fixed by not evicting a line immediately when a client request
arrives, and not fixing the first message to send. If we have the line
in cache and can't handle the request immediately or we don't have the
line and can't make room without evicting a dirty line we go into the
ready state, which now just keeps the client request. (Looking for a
clean line to evict is O(n), so it may be better just to give up if
the line's not in our cache - no client request can be handled without
hitting the bus in that case anyway). In the ready state when we
actually hit startTransaction we recheck the client request and see
what our first message should be, safe in the knowledge that no
other transactions can happen before we get to the bus, if we
decide we still need it (if we don't we still have to return
a Done message and a done state, which will show up as a transaction
in the statistics).